feat: refactored import statement targeting authslice in redux
