+++ src/components/Dashboard/SignOut.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 //import { useNavigate } from 'react-router-dom';
 import { useAppDispatch } from '../../store';
-import { resetAuthState } from '../../Redux/AuthSlice';
+import { resetAuthState } from '../../Redux/Authentication/AuthSlice';
 import '../../assets/styles/dashboardStyles/SignOut.scss';
 
 const SignOut:React.FC = () => {
