+++ src/components/authComponents/SignUpForm.tsx
@@ -3,8 +3,8 @@ import TextInput from './TextInput';
 import '../../assets/styles/authenticationStyles/SignUpForm.scss';
 import { Link, useNavigate } from 'react-router-dom';
 import { useAppDispatch } from '../../store';
-import { addNewUser } from '../../Redux/AuthSlice';
-import { NewBusiness, NewUser } from '../../Redux/Authentication/initialState';
+import { addNewUser } from '../../Redux/Authentication/AuthSlice';
+import { NewBusiness, NewUser } from '../../Redux/Authentication/authInitialStates';
 import Email from './Email';
 import Password from './Password';
 import '../../assets/styles/authenticationStyles/Login.scss';
